Many many changes, still so much to do.
Some fixes and many improvements, but I want to change something really
important next time.
Cache class must become a Status for the app, with properties handled
with beans.

+++ src/diabete/dati/GlicemiaRilevata.java
@@ -14,11 +14,11 @@ import java.util.Date;
 public class GlicemiaRilevata implements java.io.Serializable {
 	
 	public final Date timestamp;
-	public final double valore;
+	public final int valore;
 
-	public GlicemiaRilevata(Date timestamp, double valore) {
-		this.timestamp = timestamp;
-		this.valore = valore;
+	public GlicemiaRilevata(Date timestamp, int valore) {
+            this.timestamp = timestamp;
+            this.valore = valore;
 	}
 	
 }
